fix: check existence of casa files

+++ clustermgr/tasks/server.py
@@ -846,7 +846,7 @@ def install_gluu_server(task_id, server_id):
         if not as_boolean(setup_prop['oxd_use_gluu_storage']):
             #we need to copy casa.json and oxd_db.mv.db from primry server
             casa_files = [
-                        #os.path.join(installer.container, 'opt/oxd-server/data/oxd_db.mv.db')
+                        os.path.join(installer.container, 'opt/oxd-server/data/oxd_db.mv.db')
                         ]
 
             if installer.gluu_version.replace('nochroot-', '') < '4.2.0':
@@ -856,9 +856,10 @@ def install_gluu_server(task_id, server_id):
             os.mkdir(tmp_dir)
             for remote_fp in casa_files:
                 local_fp = os.path.join(tmp_dir, os.path.basename(remote_fp))
-                primary_server_installer.download_file(remote_fp, local_fp)
-                installer.upload_file(local_fp, remote_fp)
-                installer.run('chown jetty:jetty ' + remote_fp)
+                if primary_server_installer.conn.exists(remote_fp):
+                    primary_server_installer.download_file(remote_fp, local_fp)
+                    installer.upload_file(local_fp, remote_fp)
+                    installer.run('chown jetty:jetty ' + remote_fp)
 
         wlogger.log(task_id, "Gluu Server successfully installed")
